move and rename kubernetes.go
