+++ test/e2e/envutils.go
@@ -1,7 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: MIT
 
-package common
+package e2e
 
 import (
 	"fmt"
